feat: implement admin dashboard, event management system, and Cloudinary integration with secure API authentication

+++ src/app/api/event/[id]/route.ts
@@ -1,4 +1,4 @@
-import { prisma } from "@/lib/prisma";
+import { AuthenticationError, verifyApiSession } from "@/lib/api-session";
 import { EventServices } from "@/services/event.services";
 import { EventSchema } from "@/validations/event.schema";
 import { NextRequest, NextResponse } from "next/server";
@@ -7,6 +7,7 @@ const service = new EventServices();
 
 export async function GET(request: NextRequest, { params }: { params: Promise<{ id: string }> }) {
     try {
+
         const { id } = await params;
         const event = await service.findById(id);
         if (!event) {
@@ -21,21 +22,33 @@ export async function GET(request: NextRequest, { params }: { params: Promise<{ 
 
 export async function DELETE(request: NextRequest, { params }: { params: Promise<{ id: string }> }) {
     try {
+        const session = await verifyApiSession()
+        if (session.role !== "ADMIN") {
+            return NextResponse.json({ success: false, message: "You do not have access to this feature" }, { status: 401 })
+        }
         const { id } = await params;
 
         await service.delete(id)
 
         return NextResponse.json({ message: "Event deleted successfully" })
     } catch (error) {
+        if (error instanceof AuthenticationError) {
+            return NextResponse.json({ message: error.message }, { status: 403 })
+        }
 
-        console.log(error)
-
-        return NextResponse.json({ message: "Event delete failed" }, { status: 500 })
+        return NextResponse.json(
+            { success: false, message: "Internal Server Error" },
+            { status: 500 }
+        );
     }
 }
 
 export async function PATCH(request: NextRequest, { params }: { params: Promise<{ id: string }> }) {
     try {
+        const session = await verifyApiSession()
+        if (session.role !== "ADMIN") {
+            return NextResponse.json({ success: false, message: "You do not have access to this feature" }, { status: 403 })
+        }
         const { id } = await params;
 
         const body = await request.json();
@@ -46,6 +59,9 @@ export async function PATCH(request: NextRequest, { params }: { params: Promise<
 
         return NextResponse.json(event);
     } catch (error) {
+        if (error instanceof AuthenticationError) {
+            return NextResponse.json({ message: error.message }, { status: 403 })
+        }
 
         console.error(error);
